feat(health): add template test variable function.
Add new template functions to handle CPU temperature and details,
and update the templates to use these functions.
This includes adding the isset function to check
if a key exists in a map or array,
and updating the cpu_details.tmpl and critical_metrics.tmpl templates to use these functions.